requirements.txt
flask==2.3.3
openai==1.3.7
whisper-openai==1.1.10
pytesseract==0.3.10
pillow==10.0.1
opencv-python==4.8.1.78
python-dotenv==1.0.0
langdetect==1.0.9
cryptography==41.0.4
pydantic==2.4.2
sqlite3
requests==2.31.0
pytest==7.4.2
pytest-asyncio==0.21.1
twilio==8.10.2
